Enhancements

+++ service.sh
@@ -15,16 +15,25 @@ if [ -d $DIR ] && [ ! -f $AML/disable ]; then
   chcon -R u:object_r:vendor_configs_file:s0 $DIR
 fi
 
+# magisk
+if [ -d /sbin/.magisk ]; then
+  MAGISKTMP=/sbin/.magisk
+else
+  MAGISKTMP=`realpath /dev/*/.magisk`
+fi
+
+# path
+MIRROR=$MAGISKTMP/mirror
+SYSTEM=`realpath $MIRROR/system`
+VENDOR=`realpath $MIRROR/vendor`
+ODM=`realpath $MIRROR/odm`
+MY_PRODUCT=`realpath $MIRROR/my_product`
+
 # function
 bind_other_etc() {
-NAME=*mixer*paths*.xml
-if [ ! -d $AML ] || [ -f $AML/disable ]; then
-  DIR=$MODPATH/system/vendor
-else
-  DIR=$AML/system/vendor
-fi
 FILE=`find $DIR/etc -maxdepth 1 -type f -name $NAME`
-if [ `realpath /odm/etc` == /odm/etc ] && [ "$FILE" ]; then
+if [ ! -d $ODM ] && [ "`realpath /odm/etc`" == /odm/etc ]\
+&& [ "$FILE" ]; then
   for i in $FILE; do
     j="/odm$(echo $i | sed "s|$DIR||")"
     if [ -f $j ]; then
@@ -33,7 +42,8 @@ if [ `realpath /odm/etc` == /odm/etc ] && [ "$FILE" ]; then
     fi
   done
 fi
-if [ -d /my_product/etc ] && [ "$FILE" ]; then
+if [ ! -d $MY_PRODUCT ] && [ -d /my_product/etc ]\
+&& [ "$FILE" ]; then
   for i in $FILE; do
     j="/my_product$(echo $i | sed "s|$DIR||")"
     if [ -f $j ]; then
@@ -45,13 +55,27 @@ fi
 }
 
 # mount
-#pbind_other_etc
+NAME=*mixer*paths*.xml
+if [ -d $AML ] && [ ! -f $AML/disable ]\
+&& find $AML/system/vendor -type f -name $NAME; then
+  DIR=$AML/system/vendor
+#p  bind_other_etc
+else
+  DIR=$MODPATH/system/vendor
+  bind_other_etc
+fi
 
 # restart
 if [ "$API" -ge 24 ]; then
-  killall audioserver
+  PID=`pidof audioserver`
+  if [ "$PID" ]; then
+    killall audioserver
+  fi
 else
-  killall mediaserver
+  PID=`pidof mediaserver`
+  if [ "$PID" ]; then
+    killall mediaserver
+  fi
 fi
